# fix issue 18315 - wrong code for `i > 0`