~~Removing the wrong optimization, falling back to cmp instruction.~~

----

I have no idea what I'm doing when it comes to dmd code, and this seems like a place where one can introduce subtle bugs. Please review carefully.