@timotheecour 2018-02-04
I think a benchmark showing effect on performance would be nice to know how much this affects dmd.
Since there was no bug with `-O` (and ldc2 isn't affected) my guess is shouldn't affect much.