@aG0aep6G 2018-02-05
> I think a benchmark showing effect on performance would be nice to know how much this affects dmd.

For this D code:

```d
bool f(int i) { return i > 0; }
```

dmd 2.078.0 generates this:

```asm
   0:   push   rbp
   1:   mov    rbp,rsp
   4:   mov    rax,rdi
   7:   neg    eax
   9:   shr    eax,0x1f
   c:   pop    rbp
   d:   ret
```

with this PR:

```asm
   0:   push   rbp
   1:   mov    rbp,rsp
   4:   sub    rsp,0x10
   8:   mov    DWORD PTR [rbp-0x8],edi
   b:   cmp    DWORD PTR [rbp-0x8],0x0
   f:   setg   al
  12:   mov    rsp,rbp
  15:   pop    rbp
  16:   ret
```

LDC:

```asm
   0:   test   edi,edi
   2:   setg   al
   5:   ret
```

(All with `-O`.)

I'm not sure if a fair comparison includes the prologue/epilogue in this case. Anyway, here are some times with function calls:
```bash
# Compile old version.
echo 'bool f_old(int i) { return i > 0; }' | dmd2.078 - -c -ofold.o -O -betterC
# Compile new version.
echo 'bool f_new(int i) { return i > 0; }' | dmd2.git - -c -ofnew.o -O -betterC
# Try LDC too, for reference.
echo 'bool f_ldc(int i) { return i > 0; }' | ldc2 - -c -ofldc.o -O -betterC
# Compile timing code, link, run.
dmd -O old.o new.o ldc.o -run - <<EOF
bool f_old(int);
bool f_new(int);
bool f_ldc(int);
void main()
{
    import std.conv;
    import std.datetime;
    import std.datetime.stopwatch: benchmark; /* Really? */
    import std.stdio;
    auto ts = benchmark!(
        () { f_old(int.min); },
        () { f_new(int.min); },
        () { f_ldc(int.min); },
    )(10_000_000);
    writeln("old: ", ts[0].to!Duration);
    writeln("new: ", ts[1].to!Duration);
    writeln("LDC: ", ts[2].to!Duration);
}
EOF
```

Typical output:
```
old: 28 ms, 140 μs, and 2 hnsecs
new: 31 ms, 107 μs, and 2 hnsecs
LDC: 23 ms, 451 μs, and 6 hnsecs
```

> Since there was no bug with -O (and ldc2 isn't affected) my guess is shouldn't affect much.

The bug exists with `-O`. It's just a bit harder to trigger:

```d
void main(string[] args)
{
    int i = args.length != 0 ? int.min : 1;
    bool* b = new bool(i > 0);
    if (*b) assert(false);
}
```