@wilzbach 2018-02-05
(The Travis failure can be safely ignored - we have disabled it for `master` already)