@aG0aep6G 2018-02-06
Followed @WalterBright's advice and fixed the optimization by enabling the `SBB` instruction. I've also added some comments to explain how all this works. Also moved the test to mars1.d.