@MartinNowak 2018-02-07
Why don't we just implement the `test Xdi, Xdi; setg al` solution?
https://godbolt.org/g/8gCBrV