@WalterBright 2018-02-08
> Why don't we just implement the test Xdi, Xdi; setg al solution?

Great question.

1. The fix here is simple and not risky, and it gets the bug fixed.
2. The code gen still supports processors that don't have the `setg` instruction.
3. I plan to support `setg` when I have some time to spend on it.